updated the backend to be more user friendly
further txt file added with post formats

+++ src/main/java/com/example/nft_device_blockchain/controller/UsersController.java
@@ -28,7 +28,7 @@ public class UsersController {
     }
 
     @PutMapping("/{userId}")
-    public Users updateUser(@PathVariable Long userId, @RequestBody Users user) {
+    public Users updateUser(@PathVariable int userId, @RequestBody Users user) {
         user.setUserId(userId);
         return usersRepository.save(user);
     }
