+++ src/main/java/com/example/nft_device_blockchain/data/Warranty_logsRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface Warranty_logsRepository extends JpaRepository<Warranty_logs, Integer> {
 
     @Query("SELECT w FROM Warranty_logs w WHERE w.added_by = ?1")
-    List<Warranty_logs> findByAdded_by(Long id);
+    List<Warranty_logs> findByAdded_by(int id);
 
 }
